Refactor FlumpView API.
The existing view API specifies the schema, resource name & URL as
parameters to the constructor. This is pretty good when you've got a
bunch of simple API endpoints that work pretty much the same way - they
can all share the same class, and just have these parameters changed
appropriately.

However, when you're creating a less homogeneous API and need to specify
a class specific to an endpoint this becomes a bit clumsy.  You have to
define a class, then at a later point instantiate it with some
parameters and pass it to the flump blueprint.

I figured it would be a bit nicer if the whole process could be done at
once - you could define your custom class, provide the resource name &
schema as part of this class, and use a decorator to register it on the
blueprint.  So, I've done that.

It also seemed a bit messy to have 2 APIs for creating views - the
existing API and this new one, so I've removed support for the older
one.  It's not much more verbose, and keeps it nice and clear how to go
about implementing the advanced stuff.