Make Immutable printable (Fixes #14)
The base marshmallow Validator class provides a `__repr__` function,
that expects an attribute `error`.  We didn't have this, so printing out
a field with Immutable validator was crashing.  Not something that'd
happen very often, but can definitely happen at a repl.  For now I'm
just adding this attribute as None by default.