Stop redirecting to "canonical" URLs. (#31)
When a flump view is created with a trailing slash in it's URL, flask
assumes that that URL is the canonical version of the URL.  When a
request comes in to that URL, but missing the trailing slash, flask will
return a redirect to the canonical version of the URL.

This is fine for GET requests, but means that POST requests can end up
being forwarded to the get_many endpoint as GET requests.  If you only
pay attention to whether you're getting an error status code, this can
hide the fact that you've not actually just made a successful POST.

If you provide URLs without a trailing slash instead, then only the
non-trailing slash URL will work, and the trailing slash URL will 404.
This seems better than the current situation, but still not ideal.

Thankfully werekzeug has a `strict_slashes` option that causes routes
with and without the trailing slash to match.  So, I've gone for that
for now.

Fixes #29