Extract EntityData creation into a function. (#36)
This moves the creation of EntityData namedtuples out into a function.
This should make it easier for client applications to override the
assumption that models have `.id` and `.etag` properties.  They can
simply override this function (and the etag match checking function) and
they're good to go.

Haven't updated the documentation around this because I reckon we might
change things up again sometime soon.

Also had to change the post endpoint around a bit - it was another place
that had the `attributes.id` and `attributes.etag` accessors hardcoded,
and also relied on some code in the schema that did the same.  Was
fairly trivial to update to use the new EntityData construction
function.