Refactor Flump View construction
After working for some time with flump, and building out more
complex api's we've noticed limitations in the way we
construct our views.

With the current setup we favour a lot of inheritance, which is
fine for simpler api's, however with more complex api's it becomes
more difficult to check which methods are implemented for which views,
and how the orm logic is applied.

This commit refactors the methods to be defined as a set on the view,
rather than inheriting the methods you wish to use. This should make
it easier to identify exactly which methods are defined for a view.

We also move the logic for getting entities and their counts to a new
`Fetcher` class, which should encapsulate all of the logic needed to
get entitites.

We move the logic for deleting/updating/creating entities to a new
`OrmIntegration` class, again encapsulating this logic in one place,
rather than over different classes.

The major point of this refactor is to take more of a "composition"
approach rather than an "inheritance" approach, hopefully simplifying
the logic required to build a Flump app in the process.