Allow FlumpView URL mapping to be customised. (#43)
Why do we need this change?
----

It's not uncommon to need to write an API endpoint for which GET_MANY doesn't
make sense: for example, getting the current users profile data.  There's only
one current_user, so only one response makes sense.

Under these circumstances, it'd be nice to have the GET endpoint under `/`
rather than `/<entity_id>`.

What effects does this change have?
---

This adds a (currently undocumented) MAPPING dict to the FlumpView, which should
allow us to change the specific URL flump uses for each of the individual methods
on a FlumpView.