Add logging kwarg to FlumpBlueprint
Why do we need this change?
----

We were providing a default logger, but with propagate disabled.
This was causing issues with outputting errors due to no logger
being enabled in various apps.

What effects does this change have?
----

Provide a `logging` kwargs, which if enabled sets up a default logger on
the flump view