Don't assume a GET request exists in POST implementations. (#51)
Why do we need this change?
----

I'm currently writing an endpoint that defines HTTP_METHODS to just include
POST.  This means that the flask view I'm writing has no GET method.  However,
in POST endpoints we assume that there is a GET endpoint, and use that in the
`self` link and the `Location` header.

What effects does this change have?
----

Updates the POST endpoint to check if the current view has GET in it's
HTTP_METHODS before generating the links/Location header.  `self` link is
described in the spec as a MAY, and the `Location` header is a `SHOULD`.  In
this case there's no real way to provide either, so I think we should omit.